Move eth1 http methods to execution_layer crate
