+++ beacon_node/execution_layer/Cargo.toml
@@ -16,7 +16,6 @@ reqwest = { version = "0.11.0", features = ["json","stream"] }
 eth2_serde_utils = "0.1.1"
 serde_json = "1.0.58"
 serde = { version = "1.0.116", features = ["derive"] }
-eth1 = { path = "../eth1" }
 warp = { version = "0.3.2", features = ["tls"] }
 jsonwebtoken = "8"
 environment = { path = "../../lighthouse/environment" }
