+++ beacon_node/execution_layer/src/engine_api.rs
@@ -1,7 +1,7 @@
 use crate::engines::ForkChoiceState;
 use async_trait::async_trait;
-use eth1::http::RpcError;
 pub use ethers_core::types::Transaction;
+use http::deposit_methods::RpcError;
 pub use json_structures::TransitionConfigurationV1;
 use reqwest::StatusCode;
 use serde::{Deserialize, Serialize};
